2023.05.25. Like 로직 관련 LikeDTO, LikeService, LikeRepository, LikeService, LikeController,mybatis-config, Like-Mapper 각각 생성 및 수정.

+++ src/main/java/com/icia/News/DTO/LikeDTO.java
@@ -9,9 +9,10 @@ import lombok.ToString;
 @ToString
 
 public class LikeDTO {
-    private int likeId;
-    private int articleId;
-    private int memberId;
-    private int article_like;
+    private Long likeId;
+    private Long articleId;
+    private Long memberId;
+    private Long article_like;
 
 }
+
